+++ Sources/Sentry/SentryOptions.m
@@ -46,6 +46,7 @@ SentryOptions ()
         self.sendDefaultPii = NO;
         self.enableAutoPerformanceTracking = YES;
         self.enableNetworkTracking = YES;
+        self.enableNetworkBreadcrumbs = YES;
         _defaultTracesSampleRate = nil;
         self.tracesSampleRate = _defaultTracesSampleRate;
         _experimentalEnableTraceSampling = NO;
@@ -88,8 +89,7 @@ SentryOptions ()
                       didFailWithError:(NSError *_Nullable *_Nullable)error
 {
     if (self = [self init]) {
-        [self validateOptions:options didFailWithError:error];
-        if (nil != error && nil != *error) {
+        if (![self validateOptions:options didFailWithError:error]) {
             [SentryLog
                 logWithMessage:[NSString stringWithFormat:@"Failed to initialize: %@", *error]
                       andLevel:kSentryLevelError];
@@ -115,7 +115,7 @@ SentryOptions ()
 /**
  * Populates all `SentryOptions` values from `options` dict using fallbacks/defaults if needed.
  */
-- (void)validateOptions:(NSDictionary<NSString *, id> *)options
+- (BOOL)validateOptions:(NSDictionary<NSString *, id> *)options
        didFailWithError:(NSError *_Nullable *_Nullable)error
 {
     NSPredicate *isNSString = [NSPredicate predicateWithBlock:^BOOL(
@@ -156,6 +156,9 @@ SentryOptions ()
     if ([options[@"maxBreadcrumbs"] isKindOfClass:[NSNumber class]]) {
         self.maxBreadcrumbs = [options[@"maxBreadcrumbs"] unsignedIntValue];
     }
+
+    [self setBool:options[@"enableNetworkBreadcrumbs"]
+            block:^(BOOL value) { self->_enableNetworkBreadcrumbs = value; }];
 
     if ([options[@"maxCacheItems"] isKindOfClass:[NSNumber class]]) {
         self.maxCacheItems = [options[@"maxCacheItems"] unsignedIntValue];
@@ -238,6 +241,12 @@ SentryOptions ()
 
     [self setBool:options[@"enableSwizzling"]
             block:^(BOOL value) { self->_enableSwizzling = value; }];
+
+    if (nil != error && nil != *error) {
+        return NO;
+    } else {
+        return YES;
+    }
 }
 
 - (void)setBool:(id)value block:(void (^)(BOOL))block