+++ Tests/SentryTests/Integrations/Performance/UIViewController/SentryUIViewControllerSwizzlingTests.swift
@@ -15,8 +15,8 @@ class SentryUIViewControllerSwizzlingTests: XCTestCase {
             return options
         }
         
-        var sut: SentryUIViewControllerSwizziling {
-            return SentryUIViewControllerSwizziling(options: options, dispatchQueue: TestSentryDispatchQueueWrapper())
+        var sut: SentryUIViewControllerSwizzling {
+            return SentryUIViewControllerSwizzling(options: options, dispatchQueue: TestSentryDispatchQueueWrapper())
         }
     }
     
@@ -25,6 +25,7 @@ class SentryUIViewControllerSwizzlingTests: XCTestCase {
     override func setUp() {
         super.setUp()
         fixture = Fixture()
+        SentrySDK.start(options: fixture.options)
     }
     
     override func tearDown() {
@@ -34,28 +35,31 @@ class SentryUIViewControllerSwizzlingTests: XCTestCase {
 
     func testShouldSwizzle_TestViewController() {
         let result = fixture.sut.shouldSwizzleViewController(TestViewController.self)
-
         XCTAssertTrue(result)
     }
     
     func testShouldNotSwizzle_NoImageClass() {
-        let result = fixture.sut.shouldSwizzleViewController(UIApplication.self)
+        let noImageClass: AnyClass = objc_allocateClassPair(NSObject.self, "NoImageClass", 0)!
+        let result = fixture.sut.shouldSwizzleViewController(noImageClass)
 
         XCTAssertFalse(result)
     }
     
     func testShouldNotSwizzle_UIViewController() {
         let result = fixture.sut.shouldSwizzleViewController(UIViewController.self)
-
         XCTAssertFalse(result)
     }
     
-    func testViewControllerWithoutLoadView_NoTransactionBoundToScope() {
-        let controller = TestViewController()
-        
+    func testUIViewController_loadView_noTransactionBoundToScope() {
+        let controller = UIViewController()
         controller.loadView()
-
         XCTAssertNil(SentrySDK.span)
+    }
+    
+    func testViewControllerWithoutLoadView_TransactionBoundToScope() {
+        let controller = TestViewController()
+        controller.loadView()
+        XCTAssertNotNil(SentrySDK.span)
     }
     
     func testViewControllerWithLoadView_TransactionBoundToScope() {
@@ -68,12 +72,12 @@ class SentryUIViewControllerSwizzlingTests: XCTestCase {
         XCTAssertNotNil(span)
         
         let transactionName = Dynamic(span).name.asString
-        let expectedTransactionName = SentryUIViewControllerSanitizer.sanitizeViewControllerName( controller)
+        let expectedTransactionName = SentryUIViewControllerSanitizer.sanitizeViewControllerName(controller)
         XCTAssertEqual(expectedTransactionName, transactionName)
     }
 
     func testSwizzle_fromScene() {
-        let swizzler = TestSentryUIViewControllerSwizziling(options: fixture.options, dispatchQueue: TestSentryDispatchQueueWrapper())
+        let swizzler = TestSentryUIViewControllerSwizzling(options: fixture.options, dispatchQueue: TestSentryDispatchQueueWrapper())
         
         let window = UIWindow()
         window.rootViewController = TestViewController()
@@ -88,7 +92,7 @@ class SentryUIViewControllerSwizzlingTests: XCTestCase {
     
     @available(iOS 13.0, tvOS 13.0, macCatalyst 13.0, *)
     func testSwizzle_fromScene_invalidNotification_NoObject() {
-        let swizzler = TestSentryUIViewControllerSwizziling(options: fixture.options, dispatchQueue: TestSentryDispatchQueueWrapper())
+        let swizzler = TestSentryUIViewControllerSwizzling(options: fixture.options, dispatchQueue: TestSentryDispatchQueueWrapper())
         
         let notification = Notification(name: NSNotification.Name(rawValue: "UISceneWillConnectNotification"), object: nil)
         swizzler.swizzleRootViewController(fromSceneDelegateNotification: notification)
@@ -98,7 +102,7 @@ class SentryUIViewControllerSwizzlingTests: XCTestCase {
     
     @available(iOS 13.0, tvOS 13.0, macCatalyst 13.0, *)
     func testSwizzle_fromScene_invalidNotification_ObjectNotAnArray() {
-        let swizzler = TestSentryUIViewControllerSwizziling(options: fixture.options, dispatchQueue: TestSentryDispatchQueueWrapper())
+        let swizzler = TestSentryUIViewControllerSwizzling(options: fixture.options, dispatchQueue: TestSentryDispatchQueueWrapper())
         
         let window = UIWindow()
         window.rootViewController = TestViewController()
@@ -112,7 +116,7 @@ class SentryUIViewControllerSwizzlingTests: XCTestCase {
     
     @available(iOS 13.0, tvOS 13.0, macCatalyst 13.0, *)
     func testSwizzle_fromScene_invalidNotification_WrongObjectType() {
-        let swizzler = TestSentryUIViewControllerSwizziling(options: fixture.options, dispatchQueue: TestSentryDispatchQueueWrapper())
+        let swizzler = TestSentryUIViewControllerSwizzling(options: fixture.options, dispatchQueue: TestSentryDispatchQueueWrapper())
         
         let notification = Notification(name: NSNotification.Name(rawValue: "UISceneWillConnectNotification"), object: "Other type of Object")
         swizzler.swizzleRootViewController(fromSceneDelegateNotification: notification)
@@ -122,13 +126,59 @@ class SentryUIViewControllerSwizzlingTests: XCTestCase {
     
     @available(iOS 13.0, tvOS 13.0, macCatalyst 13.0, *)
     func testSwizzle_fromScene_invalidNotification_ObjectWithWrongWindowProperty() {
-        let swizzler = TestSentryUIViewControllerSwizziling(options: fixture.options, dispatchQueue: TestSentryDispatchQueueWrapper())
+        let swizzler = TestSentryUIViewControllerSwizzling(options: fixture.options, dispatchQueue: TestSentryDispatchQueueWrapper())
         let notification = Notification(name: NSNotification.Name(rawValue: "UISceneWillConnectNotification"), object: ObjectWithWindowsProperty(resultOfWindows: "Windows property of the wrong type"))
         swizzler.swizzleRootViewController(fromSceneDelegateNotification: notification)
         
         XCTAssertEqual(swizzler.viewControllers.count, 0)
     }
     
+    func testSwizzle_fromApplication_noDelegate() {
+        XCTAssertFalse(fixture.sut.swizzleRootViewController(from: MockApplication()))
+    }
+    
+    func testSwizzle_fromApplication_noWindowMethod() {
+        XCTAssertFalse(fixture.sut.swizzleRootViewController(from: MockApplication(MockApplication.MockApplicationDelegateNoWindow())))
+    }
+    
+    func testSwizzle_fromApplication_noWindow() {
+        XCTAssertFalse(fixture.sut.swizzleRootViewController(from: MockApplication(MockApplication.MockApplicationDelegate(nil))))
+    }
+    
+    func testSwizzle_fromApplication_noRootViewController_InWindow() {
+        XCTAssertFalse(fixture.sut.swizzleRootViewController(from: MockApplication(MockApplication.MockApplicationDelegate(UIWindow()))))
+    }
+    
+    func testSwizzle_fromApplication() {
+        let window = UIWindow()
+        window.rootViewController = UIViewController()
+        let delegate = MockApplication.MockApplicationDelegate(window)
+        let app = MockApplication(delegate)
+        XCTAssertTrue(fixture.sut.swizzleRootViewController(from: app))
+    }
+    
+}
+
+class MockApplication: NSObject, SentryUIApplication {
+    class MockApplicationDelegate: NSObject, UIApplicationDelegate {
+        var window: UIWindow?
+        
+        init(_ window: UIWindow?) {
+            self.window = window
+        }
+    }
+    
+    class MockApplicationDelegateNoWindow: NSObject, UIApplicationDelegate {
+    }
+    
+    weak var delegate: UIApplicationDelegate?
+    
+    override init() {
+    }
+    
+    init(_ delegate: UIApplicationDelegate?) {
+        self.delegate = delegate
+    }
 }
 
 class ViewWithLoadViewController: UIViewController {
@@ -152,7 +202,7 @@ class ObjectWithWindowsProperty: NSObject {
     }
 }
 
-class TestSentryUIViewControllerSwizziling: SentryUIViewControllerSwizziling {
+class TestSentryUIViewControllerSwizzling: SentryUIViewControllerSwizzling {
     
     var viewControllers = [UIViewController]()